+++ engine/sound/snd_wavelet.cpp
@@ -39,17 +39,17 @@ void daub4(float b[], unsigned long n, int isign)
 	nh1=(nh=n >> 1)+1;
 	if (isign >= 0) {
 		for (i=1,j=1;j<=n-3;j+=2,i++) {
-			wksp[i]	   = C0*a(j)+C1*a(j+1)+C2*a(j+2)+C3*a(j+3);
-			wksp[i+nh] = C3*a(j)-C2*a(j+1)+C1*a(j+2)-C0*a(j+3);
+			wksp[i]	   = (float)( C0*a(j)+C1*a(j+1)+C2*a(j+2)+C3*a(j+3) );
+			wksp[i+nh] = (float)( C3*a(j)-C2*a(j+1)+C1*a(j+2)-C0*a(j+3) );
 		}
-		wksp[i   ] = C0*a(n-1)+C1*a(n)+C2*a(1)+C3*a(2);
-		wksp[i+nh] = C3*a(n-1)-C2*a(n)+C1*a(1)-C0*a(2);
+		wksp[i   ] = (float)( C0*a(n-1)+C1*a(n)+C2*a(1)+C3*a(2) );
+		wksp[i+nh] = (float)( C3*a(n-1)-C2*a(n)+C1*a(1)-C0*a(2) );
 	} else {
-		wksp[1] = C2*a(nh)+C1*a(n)+C0*a(1)+C3*a(nh1);
-		wksp[2] = C3*a(nh)-C0*a(n)+C1*a(1)-C2*a(nh1);
+		wksp[1] = (float)( C2*a(nh)+C1*a(n)+C0*a(1)+C3*a(nh1) );
+		wksp[2] = (float)( C3*a(nh)-C0*a(n)+C1*a(1)-C2*a(nh1) );
 		for (i=1,j=3;i<nh;i++) {
-			wksp[j++] = C2*a(i)+C1*a(i+nh)+C0*a(i+1)+C3*a(i+nh1);
-			wksp[j++] = C3*a(i)-C0*a(i+nh)+C1*a(i+1)-C2*a(i+nh1);
+			wksp[j++] = (float)( C2*a(i)+C1*a(i+nh)+C0*a(i+1)+C3*a(i+nh1) );
+			wksp[j++] = (float)( C3*a(i)-C0*a(i+nh)+C1*a(i+1)-C2*a(i+nh1) );
 		}
 	}
 	for (i=1;i<=n;i++) {
@@ -106,7 +106,7 @@ short MuLawDecode(byte uLaw) {
 	mantissa = (uLaw&0xf) + 16;
 	adjusted = (mantissa << (exponent +3)) - 128 - 4;
 
-	return (uLaw & 0x80)? adjusted : -adjusted;
+	return (short)( (uLaw & 0x80)? adjusted : -adjusted );
 }
 
 short mulawToShort[256];
@@ -127,7 +127,7 @@ void encodeWavelet( sfx_t *sfx, short *packets) {
 
 	if (!madeTable) {
 		for (i=0;i<256;i++) {
-			mulawToShort[i] = (float)MuLawDecode((byte)i);
+			mulawToShort[i] = (short)( (float)MuLawDecode((byte)i) );
 		}
 		madeTable = qtrue;
 	}
@@ -186,7 +186,7 @@ void decodeWavelet(sndBuffer *chunk, short *to) {
 	if (!to) return;
 
 	for(i=0; i<size; i++) {
-		to[i] = wksp[i];
+		to[i] = (short)( wksp[i] );
 	}
 }
 
@@ -198,7 +198,7 @@ void encodeMuLaw( sfx_t *sfx, short *packets) {
 
 	if (!madeTable) {
 		for (i=0;i<256;i++) {
-			mulawToShort[i] = (float)MuLawDecode((byte)i);
+			mulawToShort[i] = (short)( (float)MuLawDecode((byte)i) );
 		}
 		madeTable = qtrue;
 	}
